+++ src/backend/tests/conftest.py
@@ -1,0 +1,5 @@
+"""Module py.test fixtures."""
+
+# pylint: disable=unused-import
+
+from .fixtures.backends import es